test(integration): reorganize integration tests
Refs: #414
