--- tests/src/integration/array.rs
@@ -1,4 +1,0 @@
-#[test]
-fn array_works() {
-    assert!(crate::integration::run_php("array.php"));
-}
