--- tests/src/integration/binary.rs
@@ -1,4 +1,0 @@
-#[test]
-fn binary_works() {
-    assert!(crate::integration::run_php("binary.php"));
-}
