--- tests/src/integration/bool.php
@@ -1,6 +1,0 @@
-<?php
-
-require('_utils.php');
-
-assert(test_bool(true) === true);
-assert(test_bool(false) === false);
