--- tests/src/integration/bool.rs
@@ -1,4 +1,0 @@
-#[test]
-fn bool_works() {
-    assert!(crate::integration::run_php("bool.php"));
-}
