+++ tests/src/integration/bool/bool.php
@@ -1,0 +1,4 @@
+<?php
+
+assert(test_bool(true) === true);
+assert(test_bool(false) === false);
