+++ tests/src/integration/bool/mod.rs
@@ -1,0 +1,18 @@
+use ext_php_rs::prelude::*;
+
+#[php_function]
+pub fn test_bool(a: bool) -> bool {
+    a
+}
+
+pub fn build_module(builder: ModuleBuilder) -> ModuleBuilder {
+    builder.function(wrap_function!(test_bool))
+}
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn bool_works() {
+        assert!(crate::integration::test::run_php("bool/bool.php"));
+    }
+}
