--- tests/src/integration/callable.php
@@ -1,5 +1,0 @@
-<?php
-
-require('_utils.php');
-
-assert(test_callable(fn (string $a) => $a, 'test') === 'test');
