--- tests/src/integration/callable.rs
@@ -1,4 +1,0 @@
-#[test]
-fn callable_works() {
-    assert!(crate::integration::run_php("callable.php"));
-}
