+++ tests/src/integration/callable/callable.php
@@ -1,0 +1,3 @@
+<?php
+
+assert(test_callable(fn (string $a) => $a, 'test') === 'test');
