--- tests/src/integration/class.rs
@@ -1,4 +1,0 @@
-#[test]
-fn class_works() {
-    assert!(crate::integration::run_php("class.php"));
-}
