--- tests/src/integration/closure.rs
@@ -1,4 +1,0 @@
-#[test]
-fn closure_works() {
-    assert!(crate::integration::run_php("closure.php"));
-}
