--- tests/src/integration/globals.rs
@@ -1,4 +1,0 @@
-#[test]
-fn globals_works() {
-    assert!(crate::integration::run_php("globals.php"));
-}
