+++ tests/src/integration/globals/globals.php
