--- tests/src/integration/iterator.rs
@@ -1,4 +1,0 @@
-#[test]
-fn iterator_works() {
-    assert!(crate::integration::run_php("iterator.php"));
-}
