--- tests/src/integration/magic_method.rs
@@ -1,4 +1,0 @@
-#[test]
-fn magic_method() {
-    assert!(crate::integration::run_php("magic_method.php"));
-}
