+++ tests/src/integration/magic_method/magic_method.php
