+++ tests/src/integration/mod.rs
@@ -1,0 +1,72 @@
+pub mod array;
+pub mod binary;
+pub mod bool;
+pub mod callable;
+pub mod class;
+pub mod closure;
+pub mod globals;
+pub mod iterator;
+pub mod magic_method;
+pub mod nullable;
+pub mod number;
+pub mod object;
+pub mod string;
+pub mod types;
+pub mod variadic_args;
+
+#[cfg(test)]
+mod test {
+    use std::env;
+
+    use std::process::Command;
+    use std::sync::Once;
+
+    static BUILD: Once = Once::new();
+
+    fn setup() {
+        BUILD.call_once(|| {
+            assert!(Command::new("cargo")
+                .arg("build")
+                .output()
+                .expect("failed to build extension")
+                .status
+                .success());
+        });
+    }
+
+    pub fn run_php(file: &str) -> bool {
+        setup();
+        let mut path = env::current_dir().expect("Could not get cwd");
+        path.pop();
+        path.push("target");
+        path.push("debug");
+        path.push(if std::env::consts::DLL_EXTENSION == "dll" {
+            "tests"
+        } else {
+            "libtests"
+        });
+        path.set_extension(std::env::consts::DLL_EXTENSION);
+        let output = Command::new("php")
+            .arg(format!("-dextension={}", path.to_str().unwrap()))
+            .arg("-dassert.active=1")
+            .arg("-dassert.exception=1")
+            .arg("-dzend.assertions=1")
+            .arg(format!("src/integration/{file}"))
+            .output()
+            .expect("failed to run php file");
+        if output.status.success() {
+            true
+        } else {
+            panic!(
+                "
+                status: {}
+                stdout: {}
+                stderr: {}
+                ",
+                output.status,
+                String::from_utf8(output.stdout).unwrap(),
+                String::from_utf8(output.stderr).unwrap()
+            );
+        }
+    }
+}
