--- tests/src/integration/nullable.php
@@ -1,6 +1,0 @@
-<?php
-
-require('_utils.php');
-
-assert(is_null(test_nullable()));
-assert(!is_null(test_nullable('value')));
