--- tests/src/integration/nullable.rs
@@ -1,4 +1,0 @@
-#[test]
-fn nullable_works() {
-    assert!(crate::integration::run_php("nullable.php"));
-}
