+++ tests/src/integration/nullable/nullable.php
@@ -1,0 +1,4 @@
+<?php
+
+assert(is_null(test_nullable()));
+assert(!is_null(test_nullable('value')));
