--- tests/src/integration/number.rs
@@ -1,4 +1,0 @@
-#[test]
-fn number_works() {
-    assert!(crate::integration::run_php("number.php"));
-}
