--- tests/src/integration/object.rs
@@ -1,4 +1,0 @@
-#[test]
-fn object_works() {
-    assert!(crate::integration::run_php("object.php"));
-}
