--- tests/src/integration/string.php
@@ -1,6 +1,0 @@
-<?php
-
-require('_utils.php');
-
-assert(test_str('abc') === 'abc');
-assert(test_string('abc') === 'abc');
