--- tests/src/integration/string.rs
@@ -1,4 +1,0 @@
-#[test]
-fn string_works() {
-    assert!(crate::integration::run_php("string.php"));
-}
