+++ tests/src/integration/string/string.php
@@ -1,0 +1,4 @@
+<?php
+
+assert(test_str('abc') === 'abc');
+assert(test_string('abc') === 'abc');
