--- tests/src/integration/types.rs
@@ -1,4 +1,0 @@
-#[test]
-fn types_work() {
-    assert!(crate::integration::run_php("types.php"));
-}
