+++ tests/src/integration/types/mod.rs
@@ -1,0 +1,7 @@
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn types_work() {
+        assert!(crate::integration::test::run_php("types/types.php"));
+    }
+}
