--- tests/src/integration/variadic_args.rs
@@ -1,4 +1,0 @@
-#[test]
-fn test_variadic_args() {
-    assert!(crate::integration::run_php("variadic_args.php"));
-}
